jngl 1.0.3 compatibility
In the new version of the jngl library the function names
changed to start with a lower letter. Also Time() was renamed
to getTime() and FPS() to getFPS().
The jngl library was patched to rename the print() function
to print1() because Python does not allow a function to be named
print().
Also the function Continue() is not available any more in the library.
I commented out the call in LevelEditor.py. There was no other occurence.